Relates to LBCLASSIC-51 LBCLASSIC-59 LBCLASSIC-58
Fixed synchronization problems. The iterator() method has been removed from 
the StatusManager interface. It has been replaced by getCopyOfStatusList
which, as the name indicates, returns a copy of the statusList contained in
the StatusManager instance. This avoids concurrency problems encountered
when the developer forget to synchronize on the statusManager before iterating
on its status list. 

Added a listener interface to status manager as suggested in LBCLASSIC-59 
by Anton Tagunov. 

Support for status listeners in Joran is still missing.