Related to LBCLASSIC-63
Simplifying the Joran's logic in case of errors within actions. 
In case of a failing action, Joran skip the children of the event 
associated with the action. In practice, this was already the case. 
However, with this commit, we are formally dropping support for 
SKIP_SIBLINGS.

Test cases will follow