LBCLASSIC-67
Not not all PatternLayout instances need to have a
ThrowableInformationConverter added at the end of the converter chain
by the postCompileProcessing() method.

With this commit, post compile processing code has been moved to an
interface. By default, PatternLayouts that ship with logback have a
default post compile processor. However, it is now very easy to set
processor to null.

We now test SMTPAppender using GreenMail. Problem described in
LBCLASSIC-67 is reproduced in tests but not yet fixed.