LBCORE-17
Added support for plain user password authentication. 

Furthermore, support for both STARTTLS and SSL connections were
added. Note that STARTTLS differs from SSL in that, in STARTTLS, the
connection is non-encrypted in the beginning and only after the STARTTLS 
command is issued by the client (if the server supports it) does the 
connection switch to SSL. In SSL mode, the connection is SSL from the 
start.