fixes LBCLASSIC-74
Make ch.qos.logback.classic.Logger instances deserialize with a valid LoggerContext reference.
This is accomplished by the readResolve() method which returns a Logger
generated by org.slf4j.LoggerFactory which may or not be using the correct LoggerContext. However,
that is probably the best we can do at that stage.

Associated test case.