Add support for a suffix attribute so as to distinguish the JMXConfigurator mbean of different web-application. This feature was requested in LBCLASSIC-61.
If you have two web-apps, minimal and minimal other, the logback.xml config files
relative to jmxConfigurator would look like:

<configuration>
  <jmxConfigurator suffix="Webapp=minimal" />
  ...
</configuration>

for the "minimal" web-app, and

<configuration>
  <jmxConfigurator suffix="Webapp=minimalOther" />
  ...
</configuration>

for the other web-app.

You can choose other values for the suffix contents if you wish. For
example,

  <jmxConfigurator suffix="Fruit=banana" />

is perfectly valid.

To help with garbage collection, JMXConfigurator now will clear most if
its internal fields. This is not strictly necessary but still helps with GC.