Fixing LBCLASSIC-103
Joran is now able to assume a default type for nested components according
to rules declared in a JoranConfigurator instance. The rules are located in
an instance of the DefaultNestedComponentRegistry class. 

Thus, in many cases it will no longer be necessary to declare the class of a
component in configuration files.

Since the logback-classic and logback-access modes have their own JoranConfigurator
classes, it is now possible to register rules specific to each logback module.
For example, there is now a rule which maps the layout of an appender to a 
c.q.l.classic.PatternLayout instance for all appenders in  logback-classic and to
and instance of c.q.l.access.PatternLayout for all appenders in logback-access.

This is still ongoing and incomplete work.