Fixed http://jira.qos.ch/browse/LBCORE-98
The name of the log file nested within .zip archives was incorrectly
set to be that of the tmp file logback creates during
compression. Anders Wallgren reported and posted an appropriate patch
for this problem LBCORE-98.  The nested file is now named after the
zip archive file. Analysis showed that this was an appropriate
strategy in all the cases we considered.