- started work on http://jira.qos.ch/browse/LBCLASSIC-195
The decide() method in TurboFilter now admits a LoggingEvent as an
argument instead of LoggingEvent constituents parts

The performance of logging when no turbo filters are installed remains
unchanged. However, an extra 30 nanoseconds (on a fast machine) are
added when there are one or more registered turbo filters which
corresponds to the cost of creating a LoggingEvent and garbage
collecting it even if the event is not used.

The extra cost is deemed acceptable given the flexibility it provides.