Introduce ComponentFactory
Registering a ComponentFactory with the logback context allows one to
instantiate classes that are not a part of the same classloader as
logback itself. This, in turn, allows us to support dynamically loaded
appender (and other) implementations in an OSGi environment that may
not be available as fragments - something desirable in some framework
environments like Eclipse Virgo where adding a fragment to logback
would otherwise entail various framework configuration customizations.