Addtional configuration options for sifting appender.
This change adds 2 configurable attributes to AppenderTracker and hence
SiftingAppender.  The first is maxAppenders, which limits the number of
sub-appenders the SiftingAppender will have open at any given time.
This can help prevent resource exhaustion such as running out of file
descriptors or other resources consumed by the appenders.

The second configuration option is timeout, which specifies the time
in seconds after which a sub-appender will be closed if it has not been
written to.  Previously, this value was hard coded to 30 minutes.
This option addresses RFEs http://jira.qos.ch/browse/LOGBACK-244
and http://jira.qos.ch/browse/LOGBACK-724.

These configuration options are specified as the "maxAppenders" and
"timeout" attributes on the <sift> element of logback.xml respectively.