reworked LoggingSocketServer to create SocketServer
The LoggingSocketServer was designed to work much like the existing
SimpleSocketServer, as something to be started from a main() method.
With this change, the SocketServer is now prepared to be a component
for a JoranConfigurator, allowing a socket server to be created in
any logback configuration.  This eliminates the need for the separate
configuration file for the server.