logging in ConcurrentServerRunner now uses methods of ContextAware
ConcurrentServerRunner was delegating logging to methods defined by a
subclass.  For subclasses defined in core, these methods were simply
delegating to methods of ContextAware.  For subclasses defined elsewhere
(in which a LoggerContext might be available) logging might be handled
differently.

The logging produced by ConcurrentServerRunner is mostly intended to be
diagnostic, so for consistency (no matter which module might contain the
subclass) logging is now handled exclusively through ContextAware.  This
implies that this diagnostic logging is available only when the
LoggerContext is configured for debugging, but at least that limitation
is now consistent across subclasses of ConcurrentServerRunner.