reduce timing sensitivity in SocketAppenderTest
In the receiveMessage and receiveWithContext tests, only ten
milliseconds were allowed for the appender to dispatch the message over
the socket.  Practical experience suggests that on the build system, the
actual latency is on the order of hundreds of milliseconds.  I changed
the allowed wait time to 1000 milliseconds.  This should have little
impact on test duration, as the test will finish as soon as the latch
we're awaiting is updated.

The number of iterations for the loop in lateServerLaunch was derived
from the delay allowed for the appender latch.  The new value for the
delay resulted in an iteration count of one.  I changed this test to
simply use a fixed iteration count.