modified concrete appenders to use appropriate constructors
Mutable fields in SocketAppenderBase were previously exposed as
protected fields and modified by constructors in concrete subclasses.
This commit improves encapsulation of the base class by exposing
protected constructors for subclasses to invoke rather than directly
exposing the fields.  These new constructors are marked as deprecated,
as the convenience constructors on the base classes (which are also
deprecated) are the sole reason for their existence.  A future release
should remove all but the no-arg constructor from the base class and
subclasses.