Fix orphan tmp files from failed rollover (LOGBACK-992)
TimeBasedRollingPolicy renamed the original file to a temporary file
and then started asynchronous compression of the temporary file. The
compression method bailed if the destination file already existed,
leaving the temporary file behind as an orphan.

This patch adds a flag to unconditionally delete the file, ensuring
that the tmp file is deleted regardless of any errors.