Update TimeBasedRollingPolicy.java
There may be a problem here. 
The log configurations of two load balancing projects are exactly the same, and they are all attached to the same log. 
When the log rollover at 0'clock, there will be repeated renaming operations in Linux environment, resulting in log confusion. 
So we can add a judgment that if elapsedperiodsfile already exists, there is no need to rename it.

Thank you!