Don't use URLDecoder to decode file URLs in ConfigurationWatchList
Java's `File#toURI` method does not encode all reserved URI characters
like the `+` character, because they only need to be encoded in other
parts of the URI. The `java.net.URLDecoder` on the other hand expects
the `x-www-form-urlencode` encoding and therefore replaces these
characters. Automatic reloading of logback configuration files
in directories with the `+` did therefore not work. Fixes #950

Signed-off-by: Daniel Meier <danielmeier50@gmail.com>