ViewStatusMessagesServlet requires method POST for button 'Clear' (#971)
The button 'Clear' has a side-effect and should not work with GET,
as GET is considered a Safe Method not taking an action other than retrieval.

https://www.rfc-editor.org/rfc/rfc2616#section-9.1.1

I'd like to restrict users from doing any changes by restricting them to method GET.
With that said one might consider this change as a security fix.

Signed-off-by: Ralf Wiebicke <ralf.wiebicke@exedio.com>